Use the Pipecat-provided runner for bot-ready-signalling
Replace the hand-rolled server.py + runner.py with `pipecat.runner.run`,
matching the canonical pattern in travel-companion / daily-custom-tracks.
Server is now started via `uv run bot.py -t daily` and exposes the standard
`/start` endpoint (Pipecat Cloud compatible).

JS and RN clients call `/start` with `{createDailyRoom: true}`. The Vite dev
proxy is updated from `/connect` to `/start`. Smoke-tested locally: POST
/start returns {dailyRoom, dailyToken, sessionId} and the bot joins.

Addresses filipi87's review on PR #209.
