+++ bot-ready-signalling/client/javascript/vite.config.js
@@ -3,8 +3,8 @@ import { defineConfig } from 'vite';
 export default defineConfig({
     server: {
         proxy: {
-            // Proxy /api requests to the backend server
-            '/connect': {
+            // Proxy /start (Pipecat runner) to the backend server.
+            '/start': {
                 target: 'http://0.0.0.0:7860', // Replace with your backend URL
                 changeOrigin: true,
             },
