+++ bot-ready-signalling/server/pyproject.toml
@@ -5,9 +5,7 @@ description = "Pipecat example - server"
 requires-python = ">=3.11"
 dependencies = [
     "python-dotenv",
-    "fastapi[all]",
-    "uvicorn",
-    "pipecat-ai[daily,cartesia,openai]>=1.0.0",
+    "pipecat-ai[daily,cartesia,runner]>=1.0.0",
 ]
 
 [dependency-groups]
